atualizando classe
